+++ src/test/kotlin/snc/openchargingnetwork/node/ApplicationTest.kt
@@ -2,8 +2,10 @@ package snc.openchargingnetwork.node
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@SpringBootTest
+@ActiveProfiles("test")
 class ApplicationTest {
 
     @Test