+++ src/test/kotlin/snc/openchargingnetwork/node/components/OcpiRequestHandlerTest.kt
@@ -17,6 +17,8 @@ import snc.openchargingnetwork.node.models.entities.OcnRules
 import snc.openchargingnetwork.node.models.ocpi.*
 import snc.openchargingnetwork.node.services.*
 import snc.openchargingnetwork.node.tools.generatePrivateKey
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 
 class OcpiRequestHandlerTest {
 
@@ -28,9 +30,19 @@ class OcpiRequestHandlerTest {
     private val asyncTaskService: AsyncTaskService = mockk()
     private val properties: NodeProperties = mockk()
     private val responseHandlerBuilder: OcpiResponseHandlerBuilder = mockk()
+    private val coroutineScope = CoroutineScope(Dispatchers.Unconfined)
 
-    private val requestHandlerBuilder = OcpiRequestHandlerBuilder(routingService, registryService, httpService, walletService,
-            hubClientInfoService, asyncTaskService, responseHandlerBuilder, properties)
+    private val requestHandlerBuilder = OcpiRequestHandlerBuilder(
+        routingService,
+        registryService,
+        httpService,
+        walletService,
+        hubClientInfoService,
+        asyncTaskService,
+        responseHandlerBuilder,
+        properties,
+        coroutineScope
+    )
 
     @Test
     fun forwardRequest_local() {
@@ -74,7 +86,6 @@ class OcpiRequestHandlerTest {
         every { hubClientInfoService.renewClientConnection(variables.headers.sender) } just Runs
         every { hubClientInfoService.renewClientConnection(variables.headers.receiver) } just Runs
         every { asyncTaskService.forwardOcpiRequestToLinkedServices(requestHandler) } just Runs
-        every { registryService.getAgreementsByInterface(variables.headers.sender, variables.module, variables.interfaceRole) } returns sequenceOf()
         every { responseHandlerBuilder.build(variables, expectedResponse) } returns responseHandler
         every { responseHandler.getResponse() } returns ResponseEntity.ok(expectedResponse.body)