+++ src/test/kotlin/snc/openchargingnetwork/node/integration/parties/PartyServer.kt
@@ -1,17 +1,17 @@
 package snc.openchargingnetwork.node.integration.parties
 
+import com.olisystems.ocnregistryv2_0.OcnRegistry
 import io.javalin.Javalin
 import org.web3j.tx.ClientTransactionManager
 import shareandcharge.openchargingnetwork.notary.Notary
 import shareandcharge.openchargingnetwork.notary.SignableHeaders
 import shareandcharge.openchargingnetwork.notary.ValuesToSign
-import snc.openchargingnetwork.contracts.Permissions
-import snc.openchargingnetwork.contracts.Registry
 import snc.openchargingnetwork.node.integration.utils.*
 import snc.openchargingnetwork.node.models.OcnServicePermission
 import snc.openchargingnetwork.node.models.OcnRulesListType
 import snc.openchargingnetwork.node.models.ocpi.*
 import snc.openchargingnetwork.node.tools.generateUUIDv4Token
+import snc.openchargingnetwork.node.tools.toBs64String
 
 open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContracts) {
 
@@ -27,8 +27,7 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
     // replace deployed contract instances with own party's transaction manager
     private val txManager = ClientTransactionManager(web3, config.credentials.address)
     val contracts = deployedContracts.copy(
-            registry = Registry.load(deployedContracts.registry.contractAddress, web3, txManager, gasProvider),
-            permissions = Permissions.load(deployedContracts.permissions.contractAddress, web3, txManager, gasProvider))
+            registry = OcnRegistry.load(deployedContracts.registry.contractAddress, web3, txManager, gasProvider))
 
     init {
         app.exception(JavalinException::class.java) { e, ctx ->
@@ -36,7 +35,7 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
         }
 
         app.before {
-            if (it.header("Authorization") != "Token $tokenB") {
+            if (it.header("Authorization") != "Token ${tokenB.toBs64String()}") {
                 throw JavalinException(message = "Unauthorized")
             }
         }
@@ -61,27 +60,18 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
         return Notary().sign(valuesToSign, config.credentials.privateKey()).serialize()
     }
 
-    fun setPartyInRegistry(operator: String, role: Role) {
+    fun setPartyInRegistry(operator: String, role: Role, name: String, url: String) {
         val (id, country) = config.party
         val rolesList = listOf(role.ordinal.toBigInteger())
-        contracts.registry.setParty(country.toByteArray(), id.toByteArray(), rolesList, operator).sendAsync().get()
+        contracts.registry.setParty(country.toByteArray(), id.toByteArray(), rolesList, operator, name, url).sendAsync().get()
         node = contracts.registry.getNode(operator).sendAsync().get()
     }
 
-    fun setServicePermissions(permissions: List<OcnServicePermission>) {
-        val name = "Test Service" // optional name
-        val url = "https://test.Service"  // optional public url
-        val permissionsIntList = permissions.map { it.ordinal.toBigInteger() }
-        contracts.permissions.setService(name, url, permissionsIntList).sendAsync().get()
-    }
-
-    fun agreeToServicePermissions(provider: String) {
-        contracts.permissions.createAgreement(provider).sendAsync().get()
-    }
 
     fun registerCredentials() {
         // TODO: could also request versions and store endpoints in memory
-        val tokenA = getTokenA(node, listOf(config.party))
+        val tokenA = getTokenA(node, listOf(config.party)).toBs64String()
+
         val response = khttp.post("$node/ocpi/2.2/credentials",
                 headers = mapOf("Authorization" to "Token $tokenA"),
                 json = coerceToJson(Credentials(
@@ -92,7 +82,12 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
                                 businessDetails = BusinessDetails(name = "Some CPO"),
                                 countryCode = config.party.country,
                                 partyID = config.party.id)))))
-        tokenC = response.jsonObject.getJSONObject("data").getString("token")
+
+        tokenC = response
+            .jsonObject
+            .getJSONObject("data")
+            .getString("token")
+            .let { it.toBs64String() }
     }
 
     fun deleteCredentials() {
@@ -114,7 +109,7 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
     }
 
     fun addToList(type: OcnRulesListType, party: BasicRole, modules: List<String>? = listOf()) {
-        khttp.post("$node/ocpi/receiver/2.2/ocnrules/${type.toString().toLowerCase()}",
+        khttp.post("$node/ocpi/receiver/2.2/ocnrules/${type.toString().lowercase()}",
                 headers = mapOf("Authorization" to "Token $tokenC"),
                 json = mapOf("country_code" to party.country, "party_id" to party.id, "modules" to modules))
     }